requirements.txt
cairosvg>=2.5.0
Pillow>=9.0.0 